file-logger/requirements.txt
paho_mqtt==2.1.0
configparser==7.0.0